Refactor postlist so the channel "opens" at the new message indicator
